% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{package}
\name{arcpyrextra-package}
\alias{arcpyrextra}
\alias{arcpyrextra-package}
\title{Extensions to arcpyr}
\description{
Extends the funcitonality of \code{arcpyr}. Provides some additional
interfaces to working with arcpy from R, including support for
raster calculations and reading attribute tables.
}

