% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{da_read}
\alias{da_read}
\title{Read Table with arcpy.da}
\usage{
da_read(file, fields)
}
\arguments{
\item{fields}{A vector of field names to retreive.}

\item{layer}{The layer to access.}

\item{path}{The folder or geodatabase containing \code{layer}.}
}
\value{
a dataframe with columns corresponding to \code{fields}.
}
\description{
Read a table (e.g. attribute table of a layer) with the arcpy.da module.
}
\details{
This implementation is generally slower than accessing the
  \code{@data} slot of an object created from \code{rgdal::readOGR}.
  The advantage of \code{da_read} is that it can read 
  raster attribute tables and stand-alone tables stored in file
  file geodatabases, which \code{rgdal::readOGR} cannot.
}

