% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{da_write}
\alias{da_write}
\title{Write/Update Table with arcpy.da}
\usage{
da_write(d, file, fmt, update = FALSE)
}
\arguments{
\item{d}{The data to write to the table.}

\item{file}{The file path to the table.}

\item{fmt}{The format string to use to write values from columns in
\code{d}.}

\item{update}{If TRUE, update the table. If FALSE, add new rows to the table.}
}
\value{
The path to \code{file}.
}
\description{
Write or update a table (e.g. attribute table of a layer) with the 
arcpy.da module.
}
\details{
When \code{update = TRUE}, \code{arcpy.da.updateCursor} is 
  used instead of \code{arcpy.da.InsertCursor}. Note that the number
  of rows of \code{d} is not checked prior to updating.
}

