% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_attrtable}
\alias{get_attrtable}
\title{Get Attribute Table with arcpy.da}
\usage{
get_attrtable(layer, path, fields)
}
\arguments{
\item{layer}{The layer to access.}

\item{path}{The folder or geodatabase containing \code{layer}.}

\item{fields}{A vector of field names to retreive.}
}
\value{
a dataframe with columns corresponding to \code{fields}.
}
\description{
Get the attribute table of a layer or raster with the arcpy.da module.
}
\details{
This implementation is generally slower than accessing the
  \code{@data} slot of an object created from \code{rgdal::readOGR}.
  The advantage of \code{get_attrtable} is that it can read 
  raster attribute tables and stand-alone tables stored in file
  file geodatabases, which \code{rgdal::readOGR} cannot.
}

