% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sa_calc}
\alias{sa_calc}
\title{Raster Calculations with arcpy.sa}
\usage{
sa_calc(expressions, inrasters = list(), outrasters = list())
}
\arguments{
\item{expressions}{A named list of raster calculation expressions to 
be executed in sequence. The names must be valid python object 
names.}

\item{inrasters}{A named list of file paths to rasters to be loaded
prior to performing the calculations listed in \code{expressions}.
The names must be valid python object names. Raster objects will 
be created via \code{arcpy.sa.Raster()}.}

\item{outrasters}{A named list of file paths to save rasters 
resulting from the calculations in \code{expressions}. The names 
must be valid python object names. Rasters will be saved via
\code{<raster object>.save()}}
}
\value{
No return value; raster files listed in \code{outrasters} 
  will be saved.
}
\description{
Perform Raster Calculations with arcpy.sa.
}
\details{
Note that this function exposes the \code{arcpy.sa} 
  namespace, rather than importing all functions from 
  \code{arcpy.sa} into the global environment. Therefore, 
  calls to \code{arcpy.sa} functions in the \code{expressions} 
  argument must include the namespace, e.g. \code{arcpy.sa.Log10} 
  rather than simply \code{Log10}.
}
\examples{
\dontrun{ 
input.rasters = list(inras = "path/to/input/raster")
temp.rasters = list(
  firstras = "inras > 5",
  secondras = "arcpy.sa.Power(firstras, 2)",
  thirdras = "arcpy.sa.Log10(secondras)"
)
output.rasters = list(thirdras = "path/to/output/raster")
sa_calc(temp.rasters, input.rasters, output.rasters) 
}

}

