% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{da_insert}
\alias{da_insert}
\title{Insert Table with arcpy.da}
\usage{
da_insert(d, file, fmt)
}
\arguments{
\item{d}{The data to write to the table. Column names must match 
field names of the table to be written to.}

\item{file}{The file path to the table.}

\item{fmt}{Vector of formats for the columns in \code{d}. If missing,
the format will be automatically detected as numeric (\code{\%f}), 
integer (\code{\%d}), or string (\code{\%s}).}
}
\value{
The path to \code{file}.
}
\description{
Insert rows into a table (e.g. attribute table of a layer) with the 
arcpy.da module.
}
\details{
When \code{update = TRUE}, \code{arcpy.da.updateCursor} is 
  used instead of \code{arcpy.da.InsertCursor}. Note that the number
  of rows of \code{d} is not checked prior to updating.
}

