% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{da_insert}
\alias{da_insert}
\title{Insert Table with arcpy.da}
\usage{
da_insert(table.path, d, fmt)
}
\arguments{
\item{table.path}{The file path to the table.}

\item{d}{The data to write to \code{table.path}. Column names must 
match field names of the table.}

\item{fmt}{Vector of formats for the columns in \code{d}. If missing,
the format will be automatically detected as numeric (\code{\%f}), 
integer (\code{\%d}), or string (\code{\%s}). See \code{sprintf}
for more information on format strings.}
}
\value{
The path to the table, i.e. \code{table.path}.
}
\description{
Insert rows into a table (e.g. attribute table of a layer) with the 
arcpy.da module.
}
\examples{
\dontrun{
# create an empty table
library(arcpyr)
connect_ArcPython()
arcpy = arcpy_env(NULL, NULL)
attach_toolbox(arcpy, "management")
folder = "path/to/folder"
name = "table name"
arcpy$management$CreateTable(folder, name)
# create fields in new table
arcpy$management$AddField(file.path(folder, name), "ID", "SHORT")
arcpy$management$AddField(file.path(folder, name), "VALUE", "FLOAT")
arcpy$CreateTable_management(outpath, tablename)
# generate table contents in R
d = data.frame(ID = seq(10), VALUE = rnorm(10))
# write data to table
da_insert(file.path(folder, name), d)
}

}
\seealso{
\code{\link{sprintf}}
}

