% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{da_read}
\alias{da_read}
\title{Read Table with arcpy.da}
\usage{
da_read(table.path, fields)
}
\arguments{
\item{table.path}{The file path to the table.}

\item{fields}{A vector of field names or column indices to retrieve.}
}
\value{
a dataframe with columns corresponding to \code{fields}.
}
\description{
Read a table (e.g. attribute table of a layer) with the arcpy.da module.
}
\details{
This implementation may be faster than accessing the
  \code{@data} slot of an object created from \code{rgdal::readOGR}
  in cases where there are a very large number of features. An 
  additional advantage of \code{da_read} is that it can read 
  raster attribute tables and stand-alone tables stored in file
  geodatabases, which is not supported by \code{rgdal::readOGR}.
}
\examples{
\dontrun{
layer = "path/to/table"
fields = c("VALUE", "NOTE")
da_read(layer, fields)
}

}

