% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{RasterCalculator}
\alias{RasterCalculator}
\title{Raster Calculations with arcpy.sa}
\usage{
RasterCalculator(expressions, inrasters = list(), outrasters = list())
}
\arguments{
\item{expressions}{A list of raster calculation expressions to be 
executed in sequence.}

\item{inrasters}{A named list of file paths to Rasters used in the 
calculation expressions, to be converted to ArcPy raster objects.}

\item{outrasters}{A named list of file paths to save Rasters 
resulting the calculation expressions.}
}
\value{
No return value; raster files listed in \code{outrasters} 
  will be saved.
}
\description{
Perform Raster Calculations with arcpy.sa.
}
\details{
Note that the functions exposes the \code{arcpy.sa} 
  namespace, rather than importing all function from 
  \code{arcpy.sa} into the global environment. Therefore, 
  calls to \code{arcpy.sa} functions in the \code{expressions} 
  argument must include the namespace, e.g. \code{arcpy.sa.Log10} 
  rather than simply \code{Log10}.
}

